backend/workspace/demo_app/requirements.txt
fastapi
uvicorn
python-dotenv
stripe
firebase-admin
pydantic
